promptpilot_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
google-generativeai
pydantic-settings
supabase